共校對抽出去 #55
