+++ 語料庫/管理/檢查.py
@@ -1,4 +1,6 @@
 from 語料庫.models import 語料狀況表
+from 語料庫.models import 語料表
+from 語料庫.管理.校對 import 語料表管理
 
 
 class 檢查表(語料表):
@@ -36,4 +38,5 @@ class 檢查表管理(語料表管理):
     def save_model(self, request, obj, form, change):
         # 儲存校對者
         obj.狀況.add(語料狀況表.objects.get(狀況='愛討論'))
+        obj.檢查者 = request.user
         super(檢查表管理, self).save_model(request, obj, form, change)
